Added basic functionality to help menu and trigger system
There is now a trigger scoreboard called 'tomes' that players can address through buttons and commands and whatnot that can take numerical indices as an argument and perform actions based on that input. 1 opens splash screen, 2 the help menu, 3 the about page etc.

There is still a lot of work to be done here, but it's at least a basic framework.